Show current and new version in install dialog
This change will add a new tab to the install/update dialog that will
show the current and new version of the app.

Change-Id: Ie93ca52d779f7a0268067b14aae5c53429646830
Signed-off-by: Dirk Rettschlag <dirk.rettschlag@gmail.com>

fix next button not doing anything when upgrading package

If the package existed on the phone, you'd have to physically
swipe to the next tab before being able to install. This restores
functionality to that awesome button.

Change-Id: Idcda37b19052abd17affabb3b772b4376a240281
Signed-off-by: Roman Birg <romanbirg@gmail.com>

Change-Id: I4fce1698de3307daa5d51f3ecbb682524f5f8d74